Move gang-related test out from test_constructor_failure to avoid test timeout
